More fixes to experiments to get them to compile.
Mostly changing int to uint32_t in the types of the node members list.

+++ experiments/count_read.cpp
@@ -34,12 +34,12 @@ extern int port;
 
 int main () {
   // input number of nodes and the local node id
-  int num_nodes, node_rank;
+  uint32_t num_nodes, node_rank;
   cin >> num_nodes >> node_rank;
 
   // input the ip addresses
   map <uint32_t, string> ip_addrs;
-  for (int i = 0; i < num_nodes; ++i) {
+  for (size_t i = 0; i < num_nodes; ++i) {
     cin >> ip_addrs[i];
   }
   
@@ -54,7 +54,7 @@ int main () {
   
   // form a group with a subset of all the nodes
   vector <uint32_t> members (num_nodes);
-  for (int i = 0; i < num_nodes; ++i) {
+  for (uint32_t i = 0; i < num_nodes; ++i) {
     members[i] = i;
   }
   
@@ -73,7 +73,7 @@ int main () {
     }
   }
 
-  for (int i = 0; i < num_nodes; ++i) {
+  for (size_t i = 0; i < num_nodes; ++i) {
     if (i == node_rank) {
       continue;
     }
@@ -84,7 +84,7 @@ int main () {
   
   // the predicate
   auto f = [num_nodes] (const SST<TestRow, Mode::Reads>& sst) {
-    for (int i = 0; i < num_nodes; ++i) {
+    for (size_t i = 0; i < num_nodes; ++i) {
       if (sst[i].a < sst[LOCAL].a) {
 	return false;
       }
